more fixes from Michael Adams
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@1527 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/Files/FileHandler.h
@@ -195,7 +195,7 @@ private:
 #elif defined(SDL)
 
 	void *p;		// Pointer to resource data (malloc()ed)
-	uint32 size;	// Size of data
+	size_t size;	// Size of data
 #endif
 };
