+++ Source_Files/RenderMain/shapes_sdl.h
@@ -475,8 +475,8 @@ static bool load_collection(short collection_index, bool strip)
 			for (int j=0; j<rows; j++) {
 				int16 first = SDL_ReadBE16(p);
 				int16 last = SDL_ReadBE16(p);
-				*q++ = first >> 8; *q++ = first;
-				*q++ = last >> 8; *q++ = last;
+				*q++ = (uint8)(first >> 8); *q++ = (uint8)(first);
+				*q++ = (uint8)(last >> 8);  *q++ = (uint8)(last);
 				SDL_RWread(p, q, 1, last - first);
 				q += last - first;
 			}
